refactor: View를 views directory 하위로 이동
