+++ src/js/views/WinningNumberView.js
@@ -1,4 +1,4 @@
-import View from "../core/View.js";
+import View from "./View.js";
 
 class WinningNumbersView extends View {
   makeTemplate() {
